Move synthesis steps into Steps module
resolves #155
