+++ src/NITTA/Synthesis/Steps/ConstantFolding.hs
@@ -9,41 +9,42 @@
 {-# OPTIONS -fno-warn-orphans #-}
 
 {- |
-Module      : NITTA.Synthesis.BreakLoop
+Module      : NITTA.Synthesis.ConstantFolding
 Description :
-Copyright   : (c) Aleksandr Penskoi, 2021
+Copyright   : (c) Daniil Prohorov, 2021
 License     : BSD3
 Maintainer  : aleksandr.penskoi@gmail.com
 Stability   : experimental
 -}
-module NITTA.Synthesis.BreakLoop (
-    BreakLoopMetrics (..),
+module NITTA.Synthesis.Steps.ConstantFolding (
+    ConstantFoldingMetrics (..),
 ) where
 
 import Data.Aeson (ToJSON)
 import GHC.Generics
 import NITTA.Model.Networks.Bus
 import NITTA.Model.Problems.Refactor
-import NITTA.Model.ProcessorUnits
+import NITTA.Model.ProcessorUnits.Types
 import NITTA.Model.TargetSystem
+import NITTA.Model.Time
 import NITTA.Synthesis.Types
 
-data BreakLoopMetrics = BreakLoopMetrics
+data ConstantFoldingMetrics = ConstantFoldingMetrics
     deriving (Generic)
 
-instance ToJSON BreakLoopMetrics
+instance ToJSON ConstantFoldingMetrics
 
 instance
     (UnitTag tag, VarValTime v x t) =>
     SynthesisDecisionCls
         (SynthesisState (TargetSystem (BusNetwork tag v x t) tag v x t) tag v x t)
         (TargetSystem (BusNetwork tag v x t) tag v x t)
-        (BreakLoop v x)
-        (BreakLoop v x)
-        BreakLoopMetrics
+        (ConstantFolding v x)
+        (ConstantFolding v x)
+        ConstantFoldingMetrics
     where
-    decisions SynthesisState{sTarget} o = [(o, breakLoopDecision sTarget o)]
+    decisions SynthesisState{sTarget} o = [(o, constantFoldingDecision sTarget o)]
 
-    parameters SynthesisState{} BreakLoop{} _ = BreakLoopMetrics
+    parameters SynthesisState{} ConstantFolding{} _ = ConstantFoldingMetrics
 
-    estimate _ctx _o _d BreakLoopMetrics = 5000
+    estimate _ctx _o _d ConstantFoldingMetrics = 5050
