+++ src/NITTA/Synthesis/Steps/Refactor.hs
@@ -3,27 +3,27 @@
 {-# OPTIONS -fno-warn-orphans #-}
 
 {- |
-Module      : NITTA.Synthesis.Refactor
+Module      : NITTA.Synthesis.Steps.Refactor
 Description : Synthesis tree representation
 Copyright   : (c) Aleksandr Penskoi, 2021
 License     : BSD3
 Maintainer  : aleksandr.penskoi@gmail.com
 Stability   : experimental
 -}
-module NITTA.Synthesis.Refactor (
-    module NITTA.Synthesis.BreakLoop,
-    module NITTA.Synthesis.OptimizeAccum,
-    module NITTA.Synthesis.ResolveDeadlock,
-    module NITTA.Synthesis.ConstantFolding,
+module NITTA.Synthesis.Steps.Refactor (
+    module NITTA.Synthesis.Steps.BreakLoop,
+    module NITTA.Synthesis.Steps.OptimizeAccum,
+    module NITTA.Synthesis.Steps.ResolveDeadlock,
+    module NITTA.Synthesis.Steps.ConstantFolding,
     isRefactor,
 ) where
 
 import Data.Maybe
 import Data.Typeable
-import NITTA.Synthesis.BreakLoop
-import NITTA.Synthesis.ConstantFolding
-import NITTA.Synthesis.OptimizeAccum
-import NITTA.Synthesis.ResolveDeadlock
+import NITTA.Synthesis.Steps.BreakLoop
+import NITTA.Synthesis.Steps.ConstantFolding
+import NITTA.Synthesis.Steps.OptimizeAccum
+import NITTA.Synthesis.Steps.ResolveDeadlock
 import NITTA.Synthesis.Types
 
 isRefactor SynthesisDecision{metrics}
